installation: remove celery
* Remove Celery code since it won't be installed anymore by REANA-Commons
  (closes reanahub/reana#188).
