+++ setup.py
@@ -44,7 +44,6 @@ setup_requires = [
 ]
 
 install_requires = [
-    'celery>=4.1.0,<4.3',
     'Flask-SQLAlchemy>=2.2',
     'Flask>=0.12',
     'fs>=2.0',
